refactor: HOT 모닝 세일 기능을 Store 도메인으로 통합
- EarlybirdDealController → StoreDetailController로 통합
- EarlybirdDealService → StoreDetailService로 통합
- HotMorningSaleResponse DTO를 deal → store 패키지로 이동
- 더 명확한 도메인 구조: Store 관련 기능을 Store 도메인에 집중
- 기존 deal 패키지 및 관련 파일들 정리

API 엔드포인트는 여전히 /api/v1/stores/morning-sale로 동일하게 작동
